Improve aom_rb_read_uleb() and aom_wb_write_uleb()
Check for and report errors in aom_rb_read_uleb().

In aom_wb_write_uleb(), change the type of the `value` input parameter
from uint64_t to uint32_t.

Change the type of the variables used with aom_rb_read_uleb() and
aom_wb_write_uleb() to uint32_t.

This merge request fixes items 1, 14, 15 in
https://gitlab.com/AOMediaCodec/avm/-/issues/933.

+++ aom_dsp/bitwriter_buffer.c
@@ -227,8 +227,8 @@ int aom_wb_count_primitive_refsubexpfin(uint16_t n, uint16_t k, int16_t ref,
 #if CONFIG_MULTILAYER_HLS
 // implementation of leb128() signaling in the specification using
 // aom_write_bit_buffer
-void aom_wb_write_uleb(struct aom_write_bit_buffer *wb, uint64_t value) {
-  uint64_t enc_val = value;
+void aom_wb_write_uleb(struct aom_write_bit_buffer *wb, uint32_t value) {
+  uint32_t enc_val = value;
   const size_t leb_size = aom_uleb_size_in_bytes(enc_val);
   for (size_t i = 0; i < leb_size; ++i) {
     uint8_t encoded_byte = enc_val & 0x7f;
