+++ aom_dsp/bitwriter_buffer.h
@@ -67,7 +67,7 @@ int aom_wb_count_primitive_refsubexpfin(uint16_t n, uint16_t k, int16_t ref,
                                         int16_t v);
 #endif  // CONFIG_LR_FRAMEFILTERS_IN_HEADER
 #if CONFIG_MULTILAYER_HLS
-void aom_wb_write_uleb(struct aom_write_bit_buffer *wb, uint64_t value);
+void aom_wb_write_uleb(struct aom_write_bit_buffer *wb, uint32_t value);
 #endif  // CONFIG_MULTILAYER_HLS
 
 #ifdef __cplusplus
